requirements.txt
Django==1.6.6
South==1.0
Unipath==1.0
django-authtools==1.0.0
psycopg2==2.5.3
wsgiref==0.1.2
